Upgrade discv5 tests
