+++ src/metrics.rs
@@ -8,11 +8,19 @@ lazy_static! {
     pub static ref METRICS: InternalMetrics = InternalMetrics::new();
 }
 
+/// A collection of metrics used throughout the server.
 pub struct InternalMetrics {
+    /// The number of active UDP sessions that are currently established.
     pub active_sessions: AtomicUsize,
+    /// The number of seconds to store received packets to taking a moving average over.
     pub moving_window: u64,
-    pub unsolicited_requests_per_second: AtomicUsize,
+    /// The number of unsolicited requests received per moving window.
+    pub unsolicited_requests_per_window: AtomicUsize,
+    /// The number of unsolicited requests per node per second taken as a moving average over the
+    /// `moving_window`.
     pub requests_per_node_per_second: RwLock<HashMap<NodeId, f64>>,
+    /// The number of unsolicited requests per IP per second taken as a moving average over the
+    /// `moving_window`.
     pub requests_per_ip_per_second: RwLock<HashMap<IpAddr, f64>>,
 }
 
@@ -21,7 +29,7 @@ impl InternalMetrics {
         InternalMetrics {
             moving_window: 5,
             active_sessions: AtomicUsize::new(0),
-            unsolicited_requests_per_second: AtomicUsize::new(0),
+            unsolicited_requests_per_window: AtomicUsize::new(0),
             requests_per_node_per_second: RwLock::new(HashMap::new()),
             requests_per_ip_per_second: RwLock::new(HashMap::new()),
         }
@@ -29,10 +37,15 @@ impl InternalMetrics {
 }
 
 #[derive(Clone, Debug)]
+/// The publicly accessible metrics that can be obtained from the Discv5 server.
 pub struct Metrics {
+    /// The number of active UDP sessions that are currently established.
     pub active_sessions: usize,
+    /// The number of unsolicited requests received per second (averaged over a moving window).
     pub unsolicited_requests_per_second: f64,
+    /// The number of unsolicited requests per node per second (averaged over a moving window).
     pub requests_per_node_per_second: HashMap<NodeId, f64>,
+    /// The number of unsolicited requests per IP per second (averaged over a moving window).
     pub requests_per_ip_per_second: HashMap<IpAddr, f64>,
 }
 
@@ -41,7 +54,7 @@ impl From<&METRICS> for Metrics {
         Metrics {
             active_sessions: internal_metrics.active_sessions.load(Ordering::Relaxed),
             unsolicited_requests_per_second: internal_metrics
-                .unsolicited_requests_per_second
+                .unsolicited_requests_per_window
                 .load(Ordering::Relaxed) as f64
                 / internal_metrics.moving_window as f64,
             requests_per_node_per_second: internal_metrics
