+++ src/service.rs
@@ -41,7 +41,8 @@ use tokio::time::Interval;
 
 mod ip_vote;
 mod query_info;
-// mod test;
+//TODO: Update service tests
+//mod test;
 
 /// The types of requests to send to the Discv5 service.
 pub enum ServiceRequest {
